+++ src/core/components/breadcrumbs/index.tsx
@@ -1,0 +1,1 @@
+export { Breadcrumbs, type IBreadcrumbsProps } from './breadcrumbs';