+++ src/modules/components/breadcrumb/index.tsx
@@ -1,1 +1,0 @@
-export { Breadcrumb, type IBreadcrumbProps } from './breadcrumb';