+++ src/modules/components/index.ts
@@ -1,6 +1,5 @@
 export * from './address';
 export * from './asset';
-export * from './breadcrumb';
 export * from './dao';
 export * from './member';
 export * from './odsModulesProvider';